Move test helpers to their own file
